ptint int

+++ ft_print_int.c
@@ -6,13 +6,13 @@
 /*   By: caking <caking@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/17 22:45:26 by caking            #+#    #+#             */
-/*   Updated: 2019/04/17 22:55:32 by caking           ###   ########.fr       */
+/*   Updated: 2019/04/17 23:14:38 by caking           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void    ft_print_int(t_printf *list, int c)
+void    ft_print_int(t_printf *list, long long int c)
 {
     ft_putnbr(c);
 }
